Test and Code: prevent dropping completed courses

+++ backend/routes/drop_course.js
@@ -32,6 +32,16 @@ router.post("", async (req, res) => {
             return res.status(400).send({ message: "Student is not registered" });
         }
 
+        let graded = await FinalGrades.findOne({ 
+            where: {
+                student_id: student.id,
+                course_code: course.course_code,
+            } });
+
+        if (graded) {
+            return res.status(400).send({ message: "Cannot drop a course with a final grade" });
+        }
+
         try {
             // delete course records
             DeliverableGrades.destroy({
